requirements.txt
django
django-filter
mysqlclient
Pillow
black